Delete GeminiExchangeAPI and process auction events as undefined
