--- src/exchanges/gemini/GeminiExchangeAPI.ts
@@ -1,25 +1,0 @@
-'use strict';
-/***************************************************************************************************************************
- * @license                                                                                                                *
- * Copyright 2017 Coinbase, Inc.                                                                                           *
- *                                                                                                                         *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance          *
- * with the License. You may obtain a copy of the License at                                                               *
- *                                                                                                                         *
- * http://www.apache.org/licenses/LICENSE-2.0                                                                              *
- *                                                                                                                         *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on     *
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the                      *
- * License for the specific language governing permissions and limitations under the License.                              *
- ***************************************************************************************************************************/
-import { Logger } from '../../utils/Logger';
-import CCXTExchangeWrapper from '../ccxt';
-
-let publicAPIInstance: CCXTExchangeWrapper;
-
-export function GeminiAPI(logger: Logger): CCXTExchangeWrapper {
-    if (!publicAPIInstance) {
-        publicAPIInstance = CCXTExchangeWrapper.createExchange('gemini', { key: process.env.GEMINI_KEY, secret: process.env.GEMINI_SECRET }, logger);
-    }
-    return publicAPIInstance;
-}
